Separate types from ./lib/platform_interface.dart
Moves all types defined in the `./lib/platform_interface.dart` file into
their own respective files. Except for adding the required `import`
statements no code has been modified.
