+++ packages/webview_flutter/webview_flutter_platform_interface/lib/src/types/types.dart
@@ -3,4 +3,8 @@
 // found in the LICENSE file.
 
 export 'auto_media_playback_policy.dart';
+export 'creation_params.dart';
 export 'javascript_mode.dart';
+export 'web_resource_error.dart';
+export 'web_resource_error_type.dart';
+export 'web_settings.dart';
