+++ packages/webview_flutter/webview_flutter_platform_interface/lib/src/webview_method_channel.dart
@@ -6,7 +6,8 @@ import 'dart:async';
 
 import 'package:flutter/services.dart';
 
-import '../platform_interface.dart';
+import 'platform_interface/platform_interface.dart';
+import 'types/types.dart';
 
 /// A [WebViewPlatformController] that uses a method channel to control the webview.
 class MethodChannelWebViewPlatform implements WebViewPlatformController {
